refactor: 테스트 파일 패키지 경로 이동
